Add chat functionality with landlord-tenant messaging
